Streamlit-Gemini-Pro-api/requirements.txt
google-generativeai==0.3.0
google-ai-generativelanguage==0.4.0